Airplane Mode in Emergency Callback Mode (ECM)
When user tries to change Airplane Mode between ON & OFF
in ECM mode, a notice should be shown to indicate it's in ECM mode,
and ask if user wants to exit ECM, if yes, then the Airplane Mode
is changed after ECM mode exit. If not, Airplane Mode should not be
changed.