Update to reflect new priorities for A2DP profile. DO NOT MERGE.
1. PRIORITY_OFF is when it is unchecked.
2. PRIORITY_ON is when it is checked.
3. PRIORITY_AUTO_CONNECT is when we want to auto connect.

When the box is checked, we want to auto connect,
so set the priority to AUTO_CONNECT.
When the user disconnects, we set the priority back
to PRIORITY_ON.

Change priority for Headset profile on the same lines as A2DP profile.

AUTO_CONNECT => we auto_connect.
ON => Settings app box is checked and we allow incoming connections.
OFF => Settings app check box is unchecked.

Dr no: Eastham
Bug: 2133530