Fix GoogleSettings check in PrivacySettings
GoogleSettingsProvider has moved to a new package as part of unbundling.
This broke the check used to decide whether to show the "Backup my
settings" preference. To make this more robust to package name changes,
I changed the code to look up the settings provider by authority
instead of looking for the package that contains it.

Fixes http://b/issue?id=2390945

Change-Id: Iaec3152faddbd938fc02f9f2dde748819c1d5cdf