Add "Power button ends call" accessibility setting.
This is part 3 of the fix for bug 2364220 "Accessibility improvements for
ending calls".

This change adds a checkbox under "Accessibility settings" to control the
new Settings.Secure.INCALL_POWER_BUTTON_BEHAVIOR value, which allows the
user to specify that the Power button should hang up while in-call
(instead of just turning off the screen.)  The checkbox is only shown on
devices that actually have a POWER button.

Yeah, it's a little strange having this under Accessibility (since it's not
that obvious *why* this feature is accessibility-related), but there's no
obvious better place.  See discussion in the bug for more info.

Bug: 2364220

Change-Id: I0fd7cf357972519b390575b9c06a4bbe46ff1c9b