Set profiles as preferred and then try to connect.
The APIs implementations are being updated to expose
A2DP and HFP APIs. A connect call is not honoured if its
not a preferred profile. So if an app changes the preference
of the profile than it should first update the state before
making the connect call. The API for updating preference
might not be exposed and will be mostly an API for app with ADMIN
privileges like Settings app.

Change-Id: I58b8e8201357bcfe615dc2322342520319d9e355