Partially revert "Conditionally remove Dock settings shortcut and Power control widget."
This reverts commit 580c80f853747741ae74fbb6f670315077586208.

So it turns out we won't have the ability to control the enabled flag from
Configuration changes because the dependency is actually in the package manager.
I'm reverting this change and going to do another one that just turns them
off for all configurations, because honeycomb is just for tablets.

Change-Id: I16b7b1e2d59a3a917448ce0361e3ce0fdd062a86