Partial fix for Bug 3495575.
Move the decrypt attempt to a AsyncTask. This will
unblock the UI thread in order for the device to
still be "responsive". There is still the issue of
decrypt taking 3+ seconds before it returns to. The
delay is still there becfore the fade but the text
field is now cleared and you can tap on keys.

Bug: 3495575
Change-Id: Icec82e83d3a09b3c0f856aa77870925fc8469625