Add "emergency call" button to CryptKeeper
This allows you to make an emergency call without needing to decrypt
your device first.

The exact appearance of the button, and the two possible icons shown to
the left of the text, are taken directly from the corresponding
framework resources (see keyguard_screen_*.xml, ic_emergency.png, and
stat_sys_phone_call.png.)

Also, the code in CryptKeeper.java for updating the state of, and
handling clicks from, the "Emergency call" button is mostly duplicated
from the corresponding code in LockPatternUtils and
LockPatternKeyguardView under frameworks/base.

Bug: 4494186
Change-Id: I36a713fdbc3281a7ba46762d47d5b61fb3cd194d