BT device picker: don't rescan after rotating the device
The Bluetooth device picker was restarting the device scan after
the screen is rotated. Fix this by only starting the scan when the
savedInstanceState Bundle passed to onCreate() is null.

Also removes "No Bluetooth devices were found nearby" from list if
the user switches to a different app and then switches back to
the device picker.

Bug: 5249380
Change-Id: I8959c850649eb713fb930ee0a8a7bcb73ca7c1aa