Fix an issue in how running services are shown.
If a process had started services, but another process was
using it for more important reasons, it would still be shown
as just with running services.  Now it will be bundled into
the other process that has a dependency on this.

This fixes an issue where something like a live wallpaper
that say sits forever with a connection to market's licensing
service would not show up to the user as using far more memory
than it should because it is keeping market running (if market
also leaves one of its services running, which it tends to do).

Change-Id: I0fd953b7221f7a3b35400007bb9e8437f09d066e