Fixed layout for "Available Devices" in Bluetooth
Bug 5646340.

'Title' and 'scanning' text are on the same line as a PreferenceCategory.
The first one has a 0dp width, compensated by a 1 weight.

When scanning is not displayed, it should be GONE rather than INVISIBLE
to let the title take the whole space.

When both are visible, they should split the space, or else, with long
texts in foreign languages, scanning takes the whole space, leaving the
title with an almost 0 width that generates a very high text because
of wrapping.

Distributing size among the two texts to accomodate for such cases.

Changed padding and alignment to get reasonnable results for Slovenian
language (i.e. very long text), when both title and scanning are displayed.

Strangely, the right scanning alignment is not enforced, although I checked
it is correctly set on the TextView. Not investigating further now since it
may not be desired by UX anyway.

Change-Id: I6b4c6fcbc350d3664749f58eee34ec0867720626