Overhaul vibrate and silent settings to match hard keys.
We go back to the old style of toggling between Mute/Vibrate/Normal.
There is an addition of a checkbox to indicate Vibrate on ring/notif.

Icons in volume dialogs are no longer touchable.

Bug: 5586083

This simplifies the logic in SoundSettings quite a bit, since there's a 1-1 mapping
with ringer mode.

Don't disable volume sliders when muted, so you can go to zero and back up by sliding
your finger. This is because the icons are not clickable anymore.

Change-Id: Ia7ce7e1bf108419eb107f03b7720101efe1598b1