Reduce the amount of I/O done on the UI thread.
There was unnecessary saving of IME settings when changing the state
of the checkboxes, causing ANRs when monkeys are switching between
Settings fragments quickly while other background activity is using
I/O as well. In fact, all I/O should be removed, but this change at least
reduces some of it.

Potential fix for:
Bug: 6414289

Change-Id: Ice96f0f9d921128b27ddbe26519a6791d9f01c75