Added head turn option for Face Unlock.
In security settings there is an option for head turn and it shows the current
status (on/off).  Clicking on this goes to another screen that describes the
feature and allows the user to it on/off.
Since the head turn option goes to a different screen instead of using a check box,
this change also rips out the rest of the code for the old liveliness checkbox.

Change-Id: I197af51084f3e751081fa159dadd6c2c80840d1e