Reduce jank in security lock selection by finishing on return
Earlier the finish of the chooser was happening before the pattern/pin chooser
was started, resulting in two transitions.

This change defers the finish to after returning from the other activity. There
is still sometimes a very small glitch on the way back, but much smaller than
the current problem.

Bug: 7342594

Change-Id: I5f4f4393b841ce31dc7236074381283155ebddbd