Fix an NPE in WifiP2pSettings UI
The NPE happens becase we save a references to the group that is to be
deleted, when the confirmation dialog is being shown to the user. We lose
that reference if orientation is changed (because the UI creates a new
WifiP2pSettings to interact with).

The fix is to save the name of the group that is being deleted in
persistent store, and re-select the group based on that.

Bug: 9004915
Change-Id: I1387a5b17fa2773748f4c8b34adddfc9516cfcf3